refactor: 기능을 기준으로 패키지 구조 정리
