+++ app/src/androidTest/java/woowacourse/movie/fixture/Fixtures.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.fixtures
+package woowacourse.movie.fixture
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
