+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MovieReservationActivity"
+            android:name=".view.reservation.MovieReservationActivity"
             android:exported="false"
-            android:parentActivityName=".MainActivity" />
+            android:parentActivityName=".view.list.MovieListActivity" />
         <activity
-            android:name=".MovieReservationCompletionActivity"
+            android:name=".view.reservation.MovieReservationCompletionActivity"
             android:exported="false"
-            android:parentActivityName=".MovieReservationActivity" />
+            android:parentActivityName=".view.reservation.MovieReservationActivity" />
         <activity
-            android:name=".MainActivity"
+            android:name=".view.list.MovieListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
