+++ app/src/main/java/woowacourse/movie/view/list/MovieAdapter.kt
@@ -1,9 +1,10 @@
-package woowacourse.movie
+package woowacourse.movie.view.list
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
+import woowacourse.movie.R
 import woowacourse.movie.domain.Movie
 import java.time.format.DateTimeFormatter
