+++ app/src/main/java/woowacourse/movie/view/list/MovieViewHolder.kt
@@ -1,9 +1,10 @@
-package woowacourse.movie
+package woowacourse.movie.view.list
 
 import android.view.View
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import woowacourse.movie.R
 
 class MovieViewHolder(
     view: View,
